ai/requirements.txt
python-dotenv
openai
datasets
sentence-transformers
scikit-learn
numpy
fastapi
uvicorn
pydantic
